[SYCL][CUDA] Refine LIT sub-group extension tests
Mark tests that fail with CUDA as XFAIL.
Tests that pass due to detection that the sub-group extension is not
supported are run normally.

The above should lead to passing test or unexpected failures the moment
work on sub-groups for CUDA would start, and therefore reminds us
automatically to revisit the LIT tests.

Signed-off-by: Bjoern Knafla <bjoern@codeplay.com>